+++ Venmo/src/main/java/com/paypal/android/venmo/VenmoFinishStartResult.kt
@@ -2,16 +2,19 @@ package com.paypal.android.venmo
 
 import com.paypal.android.corepayments.PayPalSDKError
 
-sealed class VenmoFinishStartResult() {
-    class Success(
+sealed class VenmoFinishStartResult {
+    data class Success(
         val token: String,
         val payerId: String,
         val approved: Boolean
     ) : VenmoFinishStartResult()
 
-    class Failure(val error: PayPalSDKError) : VenmoFinishStartResult()
+    data class Canceled(
+        val orderId: String? = null
+    ) : VenmoFinishStartResult()
 
-    //    class Canceled(val orderId: String?) : PayPalWebCheckoutFinishStartResult()
-//    data object NoResult : VenmoFinishStartResult()
+    data object NoResult : VenmoFinishStartResult()
+
+    data class Failure(val error: PayPalSDKError) : VenmoFinishStartResult()
 }